Added attention icon that is shown if iconified app needs attention.
Some applications change their title to indicate that there is something
that requires the user's attention. For example, a chat app might change
the title to "Chat app [n]" if there are n unread messages. Showing a
notification for every title change is not always desired, because it
might be too distracting or even annoying.

For that reason, this commit adds an optional "attention icon". It can
be set via command line argument or in the context menu. If the icon is
set and the app changes its title while being iconified, the
"attention icon" is set. As soon as the app is restored, the normal icon
is set again. Thus, this mechanism is much less pushy, but still tells
the user that there is something to look at (once).