Unregister signals to avoid deadlock during shut down
Upon receiving a signal, unregister all our custom signal handlers.

This prevents a potential deadlock while the user session is destroyed, where
we first receive a `SIGTERM` and starting shutting down, we try to undock the
application, but since X11 is shutting down, calls to X11 functions fail and
generates a `SIGHUP`, which initiates another shut down and causes a deadlock.

After this, a `SIGHUP` inside `SIGTERM` will immediately terminate the process.

Fixes: #67
Signed-off-by: Joan Bruguera <joanbrugueram@gmail.com>